Fix build issue with the wasm script
