+++ 3DViewer/viewer/index.html
@@ -1,31 +1,34 @@
 <!DOCTYPE html>
 <html>
+
 <head>
-<title>FloorspaceJS 3D Preview</title>
-<meta charset="UTF-8" />
-<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
-<script src="https://cdnjs.cloudflare.com/ajax/libs/three.js/r71/three.min.js"></script>
-<script src="https://cdn.rawgit.com/mrdoob/three.js/r71/examples/js/controls/OrbitControls.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/dat-gui/0.5/dat.gui.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/1.16.1/TweenLite.min.js"></script>
+  <title>FloorspaceJS 3D Preview</title>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/three.js/r71/three.min.js"></script>
+  <script src="https://cdn.rawgit.com/mrdoob/three.js/r71/examples/js/controls/OrbitControls.js"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/dat-gui/0.5/dat.gui.min.js"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/1.16.1/TweenLite.min.js"></script>
 </head>
+
 <body>
-<script type='text/javascript'>
+  <script type='text/javascript'>
 
-var Module = {
-  preRun: [],
-  postRun: [],
-  canvas: (function() {
-    var canvas = document.getElementById('canvas');
-    return canvas;
-  })(),
-  totalDependencies: 0,
-  monitorRunDependencies: function(left) {
-    this.totalDependencies = Math.max(this.totalDependencies, left);
-  }
-};
+    var Module = {
+      preRun: [],
+      postRun: [],
+      canvas: (function () {
+        var canvas = document.getElementById('canvas');
+        return canvas;
+      })(),
+      totalDependencies: 0,
+      monitorRunDependencies: function (left) {
+        this.totalDependencies = Math.max(this.totalDependencies, left);
+      }
+    };
 
-</script>
-<script src='./libtest_lib.js'></script>
+  </script>
+  <script src='./libtest_lib.js'></script>
 </body>
+
 </html>
