+++ index.html
@@ -12,7 +12,6 @@
     <link rel="mask-icon" href="icons/safari-pinned-tab.svg" color="#2d89ef">
     <meta name="msapplication-config" content="browserconfig.xml" />
     <title>FloorspaceJS</title>
-    <script src="3DViewer/libtest_lib.js"></script>
   </head>
   <body>
     <div id="app"></div>
